Add gitignore, Add package.json, add react packages and move TodoItem into its own file
