+++ package.json
@@ -1,0 +1,28 @@
+{
+  "name": "my-todo-app",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "dev": "parcel ./src/index.html",
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/tian-yi/my-todo-app.git"
+  },
+  "keywords": [],
+  "author": "Tianyi Wang <dev@tianyi.dev> (https://tianyi.dev)",
+  "license": "MIT",
+  "bugs": {
+    "url": "https://github.com/tian-yi/my-todo-app/issues"
+  },
+  "homepage": "https://github.com/tian-yi/my-todo-app#readme",
+  "dependencies": {
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1"
+  },
+  "devDependencies": {
+    "parcel-bundler": "^1.12.4"
+  }
+}
