+++ src/index.html
@@ -8,8 +8,6 @@
 <body>
   <div id="root">
   </div>
-  <script src="https://unpkg.com/react@16.13.1/umd/react.development.js"></script>
-  <script src="https://unpkg.com/react-dom@16.13.1/umd/react-dom.development.js"></script>
   <script src="./App.js"></script>
 </body>
 </html>
